Refactor image processing
Changed the way data is loaded to work in memory instead of creating new folders. --NOTE: this requires a "./data/brain_tumor_dataset/yes" and "./data/brain_tumor_dataset/no" path from where this file is downloaded. 

Added train, test, and validation splits on loaded data.

Updated visualization of data for clarity and to display that shuffling and splitting was successful.